scraper/requirements.txt
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
web3>=6.11.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
xlsxwriter>=3.1.0
eth-abi>=4.2.0
eth-utils>=2.3.0
urllib3>=2.0.0
certifi>=2023.0.0